MISSION_SKIP will now automatically unload units from transports if the mission comes with a not-trival MISSIONAI. (This is to prevent units from trying to guard cities while still in their transport.) Disabled a bunch of BBAI code in continueMission which did some ad-hoc hackery for picking up strandard units. My new AI_handleStranded function doesn't require the old BBAI code, and the old code allowed the AI the cheat by allowing units to step onto a boat without needing movement points.
Moved the AI_load condition in AI_attackCityMove to be above the group condition, so that units don't leave their transports to group with their old stack!
In all of the naval Move functions: changed the city-under-threat condition to not include the case where there is no defence and not attack. (This is to allow boats to actually leave from undefended cities.)
CvUnitAI::AI_load, will now only split the group trying to load if the target transport is more than 1 turn away.

Adjusted pathCost to discourage units from ending their turn on tiles that they cannot fight on. (eg. boats in cities)
Added a new function, CvSelectionGroup::cargoSpaceAvailable. (currently unused)
Strengthed the condition for the AI automatically give a skip order to units on transports.