Fixed a couple of corporation spread AI problems.
Fixed a bug in the executive AI which caused them to evaluate corporations for themselves rather than for the player they are spreading to.
The AI will now take into account the cost of removing an old corp when spreading a new corp.
Adjusted some of the weights used for automated cities in AI_bestUnit.