Fixed a bug in the iFoodBonus count in CvGame::normalizeAddFoodBonuses.
This bug essentially negated my bugfix for the plains cow problem.

Previously, I was using CvPlot::calculateMaxYield to determine which
land bonuses should count as high food plots. But I discovered today
that calculateMaxYield does not work as I expected.

I've rewritten the section in CvGame::normalizeAddFoodBonuses to avoid
using CvPlot::calculateMaxYield. And also, I've commented out the code
for CvPlot::calculateMaxYield so that it isn't mistakenly used in the
future before it is fixed.