Completely rewrote the attitude cache system.
Cache is now recalculated on the spot when something happens to
invalidate it.

CvPlayerAI::AI_invalidateAttitudeCache has been removed and repalced
with CvPlayerAI::AI_updateAttitudeCache.

The original system (from the 'CAR' mod) was hideously prone to OOS
bugs. I'd fixed many of the bugs in the past, but there were still more,
and it was extremely difficult to write robust code using that system.
(Here a couple of lingering bug examples from the old system.
CvTeam::declareWar didn't clear cache for the friends of the war target.
CvTeam::signDefensivePact didn't clear cache for the rivals of each
team. The crux of the problem is that in the CAR system, it was the
programmers responsibility to know exactly how and when attitude could
be affected in any way, and then clear the cache when it was affected.
If the programmer overlooked anything, it could cause OOS.)

The new system is more rigid in when the cache is recalculated; and so
similar 'bugs' will now no longer result in OOS. At worst they will
result in out-of-date attitude ratings. In fact, this is such a minor
effect that I've deliberately removed the cache update from a few places
where I think it is unimportant.