Changed the close-border attitude cache system.
The original close-border attitude cache system was prone to the same
OOS problems as the CAR mod cache system. So I have rewritten it using
the same, more robust, style as the new cache system.

ie. The cache is updated rigidly at the start of each turn rather than
being updated when it is first called upon.