Fixed a problem in the way replay info handled player color.
Changed CvReplayInfo::createInfo so that it does not use
CvPlayer::getPlayerColor, but rather just accesses the colour directly
from the game core to bypass the colour concealment check.

Also, changed concealUnknownCivs so that it returns false when there is
no active team. This should cover cases where the game isn't full
initialized, as well as pitboss hosts.