CvTeam::doTurn is now always called at the start of the team's turn.
Previously it was called at the end of the team's turn in
non-simultaneous turns mode. But it occurs to me that if it is called at
the end of the turn, then counters such as the counter-espionage turn
counter will not work as expected. Consider, for example,
counter-espionage for 1 turn. If doTurn is called at the end of the
turn, that 1 turn of counter-espionage will end before the other
player's turn even starts.