Rewrote CvMap::updateCenterUnit to only update plots in a certain range.
Previously, CvMap::updateCenterUnit updated the center unit for every
plot on the map. This is quite an expensive thing to do; and it caused a
noticeable delay when selecting units on a big map. The new version only
updates the plots which could potentially be in movement range of the
units being selected. The range in determined using a new static
function: KmodPathFinder::MinimumStepCost.