Cleaned up some minor problems in the use of the pathfinders.
The AI pathfinder and the UI pathfinder are now reset when a unit is promoted.
CvSelectionGroupAI::AI_update now uses the correct pathfinder reset function.

Removed CvSelectionGroup::resetPath and its python counterpart. (The
function was no longer used anywhere, and it's existence is a bit
confusing, so I removed it.)