Fixed minor problems with nuke event messages and city capture sounds.
The city capture sound is now played when the city is actually captured,
rather than when the city capture popup is displayed. This prevents the
sound from repeating if the player chooses to examine the city.

Nuke messages (eg. player X's nuke exploded, or player Y intercepted X's
nuke) are now only shown to civs who player who have met the
civilzations involved in the event.