Added UI option: Simple Selection, and changed the default behaviour.
The previous selection behaviour with the shift key is now called
'Simple [Shift] Unit Selection'. It can be enabled from the BUG options.
The non-'simple' behaviour is that when using shift to add units (not
remove), if the current selection mirrors a group, a shift will group
select, and form a group; otherwise, shift will not group-select nor
form a group.