Slightly amplified the effect scaling in AI_warCommitmentCost.
iCommitmentPerMil is now adjusted by an absolute value at the end of the
calculation. The effect of this is that the adjustments to
iCommitmentPerMil now have relatively stronger effects. (eg. relative
power, relative production, and so on, now count for a bit more than
before.) -- Also, the standard iCommitmentPerMil has been reduced.