Futher changes to foundValue, particularly wrt special food and features
iSpecialFoodPlus is now counted for all plots above the food threshold, regardless of whether they are water, or have a feature, or a bonus or whatever.
Moved and adjusted some of the feature parts of the evaluation, to fix a few minor problems / inconsistencies.
Changed the way yield is counted on the city plot to better reflect the actual game mechanics.
The bAmbitious bonus for bonuses is now applied for the city plot.