Many bug fixes and consistency fixes related to vassals and war.
Added new function, CvTeam::getMasterTeam. For vassals, this returns the
master's team id; otherwise it returns the team's own ID.

Fixed a problem in AI_warCommitmentCost, AI_startWarVal, and
getDefensivePower, which caused defensive pacts with the target team's
master to be ignored.

Corrected some denial types in AI_vassalTrade. It no longer checks for
PEACE_NOT_POSSIBLE_US; and instead checks for WAR_NOT_POSSIBLE_US. Also,
it will now return ATTITUDE_THEM if the would-be master would refuse to
plan war, or be bought into a war, against teams that the would-be
vassal is at war with.

Restored the check for PEACE_NOT_POSSIBLE_US in AI_surrenderTrade, with
the condition that it is not checked for peace-vassal deals.

When considering starting a war with a vassal, the AI will now use the
maximum of AI_noWarAttitudeProb of the target and of the master.
(Previously it only used the target's.)

AI_warCommitmentCost now uses the target team's master for most of the
calculations, including best units, productivity estimates. Also, it now
counts war weariness due to vassals and siblings of the target, as well
as from the target itself.