Fixed the condition checks in CvPlayer::isValidTriggerCorporation.
Previously, the function would always return false if a particular city
was being checked; and it would bypass the headquarters check if no
particular city was being checked.

This fixes a bug which allowed the 'corporate expansion' quest and the
'hostile takeover' quest to be assigned to players who do not control
the headquarters.