Made some adjustments to the AI's usage of nukes.
The AI will now count cargo when considering the value of nuking enemy units. (ie. the AI now cheats to see the cargo, so that it can get a better nuke target evaluation.)
Increased the AI_nukeWeight boost for civs with the manhattan project.
The civilian target weight for ICBMs is now higher against civs which have used nukes against us.