Fixed an flag update problem when switching in and out of debug mode.
The unit flags on non-visible plots were not being updated when
switching in out out of debug mode. When debug mode was turned on, there
were missing flags; when debug mode was turned off, there were lingering
flags.

This bug was a side-effect of my changes to CvMap::updateCenterUnit.

The 'debug' center unit now determines the flag.