Added a sytem to stop the AI from changing target cities too often.
New data: CvPlayerAI::m_iCityTargetTimer.
New functions: AI_getCityTargetTimer, AI_setCityTargetTimer.

When the AI is requested to target a city, the timer is set to 10 turns.
When the city targets are chosen normally, the timer is set to 4 turns.

Cities targets are still updated at random times, but the random updates
are blocked if the target timer is non-zero.