Fixed a GPP evaluation bug; adjusted yield value for perm. specialists
There was bug which caused the scaling formula to always treat the
current city as having 0 GPP per turn, which completely messed up the
scaling relative to the 'best' city.

The special yield value modifier effect in AI_permanentSpecialistValue
has been diluted, because that modifier is intended as more of a
short-term thing.