Made some corrections and adjustments to CvCityAI::AI_yieldValue
Extra points for productivity when below par are now only counted if the
player doesn't have any hurry types enabled.

When estimating how much unhappiness will wear-off before the city
grows, the estimated time for growth is now more 'optimistic' (which
means the growth time is shorter, which means the unhappiness is
greater...)

Fixed some mistakes in iGrowthValue and iDevalueRate with bFillingBar.
(The value was too high.) Also, tweaked the devalue rate for the default
case.

Slightly reduced the value of food in CvCityAI::AI_growthValuePerFood.