Removed BBAI function CvTeam::getCurrentMasterPower. Tidied some code.
CvTeam::getCurrentMasterPower was typcially used like this:
    int iPower = isAVassal() ? getCurrentMasterPower() : getPower();
This is cumbersome, and ineffecient because it has to do a vassal check
on all teams for isAVassal, and for getCurrentMasterPower.

The following code gives the same result:
    int iPower = GET_TEAM(getMasterTeam()).getPower();

Also, there could be some confusion between 'getCurrentMasterPower'
and 'getMasterPower', which are completely different functions with
completely different purposes.

So I've deleted CvTeam::getCurrentMasterPower.