Added an avenue for the AI to use to free itself as a vassal.
CvPlayerAI::AI_considerOffer now checks the for denial of existing
vassal deals and returns 'false' if the deal would be denied.

Before this change, AI vassals would never free themselves!

(Note: on closer examination, in the original BtS code, vassals could
free themselves from AI masters but not from human masters, because of
an inconsistence in the way existing deals were being checked. So when I
changed the code to be consistent, I had unintentially taken away the
AI's avenue for canceling vassal deals with anyone at all. Now it is
fixed and better than ever.)

Also, when considering capitualion, the AI will no longer be quite so
afraid of the new master's other enemies. (ie. if the would-be master is
much stronger than the other teams they are at war with, then the
would-be vassal will not use 'we're afraid of your enemies' as an excuse
to refuse capituation.)