Made some change to how city names are choosen.
When a new city is created, and a name is being chosen from the
civilization's city name list, the search will now skip a name of names
corresponding to the number of cities the player has already founded.
This means that if a player renames a city that they have founded, the
game won't immediately suggest the same previous name for the next city.

Note 1: this only applies to names taken from the civilization city names
list. It does not apply to names on the player-specific city list which
may be used by some mods.
Note 2: previously used names will again
be suggested if the civ's city name list is otherwise exhausted.

Also, if the civilization's list of city names is exhausted, then the
game will suggest a _random_ name from a random civ; rather than the
first available city name from a random civ.