Made some adjustments to the AI's decision making for chopping forests.
The yield produced from chopping is no longer counted in the evaluation
of improvements. This allows us to decide whether or not the new
improvement would be better regardless of potential immediate benefits.

The bChop checks will now be bypassed if the city is building a process.

The value of chopping is now scaled up a bit if the player hasn't yet
reached the target number of cities for the map size.