Made some AI changes in CvCityAI::AI_updateSpecialYieldMultiplier.
The adjustments associated with workboats, workers, and settlers are now
all the same. The effect of building a workboat is now less significant.

GetBetterUnit and EconomyFocus now result in a boost to commerce value.
The 'production' strategy now boosts production and lowers commerce.

Reduced the magnitude of the commerce value reduction when at war; and
made it more conditional. Also, increased the productivity value boost
for war when using dagger or crush.