Made some stop-gap adjustments to the AI's improvement yield evaluations
There are some serious flaws in the AI's evaluation of yield when
choosing which improvements to build. In particular, the value of food
does not scale when the value of jobs that could be worked with more
food; and the value of yield in general is inconistent with the way the
AI chooses which plots to work.. And so sometimes the AI will replace
good improvements with new ones which won't be worked; or replace high
food improvements with other improvements which won't have enough food
to be worked...

I don't address the core of these problem in this minor update, but I
have made a few changes with might help. In particular, I've decreased
the multiplier used for production, so that production doesn't end up
being worth signficantly more than food.