AI_getYieldWithBuild now counts yield from existing improvements.
Previously AI_getYieldWithBuild would only count yield from improvements
created by the build itself. And so, for example, if you were building a
railroad on a mine, the functino would count the natural yield of the
plot, and the extra yield from the railroad, but not from the mine
itself.

This is now corrected.