Fixed a possible crash in AI_warCommitmentCost and changed overseas cost
AI_warCommitmentCost would previously divide by zero when evaluating an
overseas war in which neither was able to build any boats. This has been
corrected.

Also, the flat adjustement of 5/4 for total war overseas has been
replaced by a factor which depends on the relative strength of the
ground forces.