The CvCity::findXRank functions now update rank for all cities at once.
This includes findPopulationRank, findBaseYieldRateRank,
findYieldRateRank, and findCommerceRateRank.

In the original BtS code, these functions would compare the current city
to every other city to find its rank, and thus only set the cache for
the city being queried. The new code gathers all of the city data, sorts
it, and set the rank cache for every city in one pass.