Rewrote CvCityAI::AI_getYieldMultipliers; fixed a food scaling problem.
The new version of AI_getYieldMultipliers is similar to old, but it is a
bit simplier, it includes a few flavour adjustments; but most
importantly, the food value multiplier now scales up as other yields
become more valued. This is to avoid the problem where production can
end up being worth more than the food that is necessary to work the
production.

Some other related CvCityAI functions have been tweaked to be consistent
with the new version of AI_getYieldMultipliers.