Added 'high food' check to bonuses added by normalizeAddFoodBonuses
Although existing land food bonuses which only net 3 food are only
counted for 2 points; bonuses which were added to boost the total score
were previously not subject to this test... This means it was still
possible to have a low food capital city in some rare situations.

This is now fixed.