ctrl/alt select will no long ignore moves when unit is already selected.
When selecting a group of units by holding ctrl/alt, if the clicked unit
is already selected then only units with moves will be selected in the
new group.

This new condition is because we can't tell the difference between a
click on the map and a click on the unit icon; and because of this, the
'ignore moves' rule can make it a bit difficult to select moveable units
when a unit without moves is already selected.