Increased the min elapsed turns for the circumnavigation bonus.
The circumnavigation bonus is only granted if the circumnavigation does
not occur on the first turn of the game. This is so that games which
start with a complete map don't give an unfair advantage to player #1.

Since CvTeam::doTurn is now called after the game turn is incremented,
the circumnatigation min-turn needed to be increased from > 0 to > 1.