CvPlayer::addSelectionGroup now also adds the group the the cycle list.
If a selection a non-empty selection group is not in the group cycle
list, it can cause problems. In particular, it can cause the game to
hang.

There are a bunch of places in the code which will add new groups the
cycle list in particular situations... but every so often, I change
something which may create some new fringe case or something which
causes the group not to be added to the list. So now, I'm cutting the
problem off at the source. Every time a new group is created, it is
immediately added to the cycle list. It's particular position in the
cycle list can be updated at a later time.