Fixed some bugs which sometimes caused confusion in the sea-settler AI.
The found values for unrevelead plots is not calculated; but there was
some old AI code which sometimes made use of those found value anyway,
which could cause the AI to target very bad settling sites.

The AI will now only consider landing in areas it has explored and
identified at least one city site in. Also, the found value of
unrevealed plots is now set to zero - so that junk values can't be
accidentally used by other bits of the AI.