Edited CvSelectionGroup::splitGroup to give a more balanced division.
splitGroup was writen to give proportional split for each unitAI type,
but it was failing to give a proportional division of the unit types
within each AI type.

I've changed it to not simply take the first units available for each AI
type. This should result in a more balanced split.

(This change should help reduce the likelihood of the AI launching a
naval assault with too many siege units.)