Made some adjustments to the AI's bombard target when evaluating units
The bombard target affects the unit evaluation when choosing which unit
to train for UNITAI_ATTACK_CITY.

This target now depends on the sum of both 'attack' and 'attack city'
units (rather than just 'attack city'); it is no longer increased for
the 'crush' strategy; it is now decreased if all of the (potential)
fighting is overseas; and the total value of the bombard rate is no
longer increased for units which ignore building defences. (If the siege
unit ignore building defences, then the non-siege units probably do as
well, and so it isn't really of any additional value.)

This bombard evaluation may need some further tweaking after this
change.