Reworked the tech-cost modifier for team games.
Tech cost is now directly proportional to the number of civs in the
team, but there is now a global tech cost reduction for games which
start with fewer than the default number of teams for the given world
size.

The combination of these two changes should give roughly the same tech
costs as before for games with evenly sized teams, but for games with
uneven teams the cost will be higher for the bigger teams.