Redesigned the evaluation of attack_city units.
The previous system was focus around a somewhat artibrary global target
bombard rate. The value of bombard units was boosted significantly while
below the target. Unfortunately, if siege units were powerful enough to
be chosen even when above the bombard target, then the AI would end up
with too many siege units - and also, the calculation of total bombard
was only a rough approximation, distrorted by the presence of defensive
units with bombard.

The new system does not use a bombard target, and instead focuses on the
number of units with a combat-limit of less than 100. The value of the
unit is reduced if there are already too many combat-limited units. The
reduction in value primarily targets the bombard component and the
collateral damage component.

The new system is still a bit of an ugly kludge, but I think it will
have fewer systematic weaknesses than the old system.