Reverted the K-Mod changes to team tech costs; and fixed a related bug.
The tech costs changes from 49a08d907ed have been reverted. (Tech costs
are now the same as they were originally.)

shareCounters no longer combines research progress of unfinished techs;
because this could cause the research to complete prematurely. The
current implementation is temperary. It's similar to the original code,
but with some scaling to fix an consistency problem when team with
uneven tech costs merge.

Note: the trade value of techs still does not depend on team size.