Fixed tech progress and some other minor issues with CvTeam::addTeam
Tech progress now carries over proportionally rather than absolutely
when a permanent alliance is formed.

Moved the espionage sharing from CvTeam::shareCounters to
CvTeam::shareItems. (Part of what I had added to shareCounters already
existed in shareItems - so I've removed the redundant code, added the
missing part to shareItems.)

Fixed a bts bug in addTeam which caused isCapitulated to not be set
correctly for 'this' team when eTeam is a vassal.