Made some adjustments to the map script. (Not Too Big or Small)
Zones now have a chance of having a buffer at the north and south poles
(to reduce the chance of ice-continents)

Single zone maps are no longer forbidden. (Althought they are still
impossible on 'large' size and larger.)

Increased the range of allowable zone sizes. The range now also scales
with map size, to allow fairly large zones on large maps, but to keep
the probability of single-zone maps low.

The number of "tiny islands" now depends on the number of zones, so that
bigger maps get more islands.