Made some further adjustments to AI_unitValue.
Increased the drop-rate for the 'siege value' of attack_city units.
Tweaked the value of some other units, particular reserve units.

Note: there are still a few places where value is subtracted due to
positive traits (such as city-attack bonus on unitai_collateral).
Ultimately all these things should be replaced with just positive values
for positive traits, and negative value for negative traits - weighted
accounting to the unitai type. The biggest missing ingredient to
achieving this is a fast and easy way to analysize the appropriateness
of the promotions available to the unit.