Implemented some basic logging of combat odds and outcomes.
If 'LOG_COMBAT_OUTCOMES' is defined, the game will log all of the
calculated combat odds and outcomes for every battle in the game. (In
CvUnit::resolveCombat)

The purpose of this combat log is to help verify the accuracy of the
combat odds calculations. This option should only be enabled when we are
explicitly trying to gather data.