Changed the conditions on some city advisor popups.
The 'liberate city' recommendation requires that the city be at least 10
turns old.

Health buildings will only be recommended if the city has more than 1
spare happiness.

Gold and Research recommendations will only be given if the city is one
of the top 1/2 for that commerce type, or top 1/3 for commerce in
general.