Made some more tweaks to CvPlayerAI::AI_espionageVal.
Increased the value of sabotaging projects.
Reduced the value of spreading unhappiness.
Reduced the value of sabotaging buildings.

Disabled sabotage completely when the city is in disorder (not for any
strategy reason, but just because the current evaluation may give wildly
inaccurate results if the city is in disorder - sometimes too high, and
sometimes too low.)