Fixed a culture problem which caused the AI to waste great engineers.
The value bonus for the initial city culture is no longer applied to
wonders or special buildings. I didn't think this would matter, because
there are other mechanisms for blocking the construction of wonders in
these situations; but I recently discovered that those mechanisms are
bypassed when the AI is deciding whether to rush a wonder with a great
engineer...

Also, reduced the priority factor for first-culture buldings.