Changed the way building flavour is used in AI_buildingValue, again.
The flavour modifier now scales as he minimum of the leader's flavour
value and the building's flavour value, rather than the product of those
two values.

The effect of a 10-10 flavour match has been increased from +66% to
+100%.

Building flavour is now ignored for sabortage evaluations and for
negative cost buildings such as academies.

bRemove is now defined at the top of the function (because it's used for
flavour as well as for defence buildings).