Changed the odds formula for short-circuit 2&3 in AI_chooseProduction.
The new formula is more sensitive to iBuildUnitProb. My intention is
that this should put a bit more of the leader's flavour back into the
number of units built, and also it should slightly decrease the rate of
military cities getting side-tracked by other infrastructure.

Overall, short-circuit 2 will be slightly less common than before.

The threshold for short-circuit 3 has been raised, so that it will catch
fewer buildings.