CvPlayer::canTrain team/national limits are now skipped if bTestVisible
The global unit limits are enforced based on the number of units ever
built, but team limits and national limits are based on the number of
units currently active. So even though the units may be currently maxed
out, they won't necesearily be like that forever - and therefore they
should still be visible.