Research is now completely halted while in anarchy.
Previously, research would continue as normal while in anarchy. The
commerce rate would be zero, but the civ would still get the 'base
research' points (which is set to 1 in the default xml). Now, research
is completely halted during anarchy, including the free base research.

The primary reason for this change is not really a good one... the
reason is that the AI cannot evaluate techs accurate while in anarchy,
and so I don't want them to have to choose their research while in
anarchy.

I've also changed a couple of things related to the triggering of
choose-research popups.