Made some adjustments to CvGameTextMgr::buildBuildingRequiresString
For buildings such as catherals which require a number of other
buildings nation-wide, the requirement count is now includes the number
of (catheral style) buildings currently under construction. This means
if a catheral cannot be added to the build queue of multiple cities due
to a lack of temples, the city screen will now accurately report the
number of temples required.

Also, in the civilopedia such requirement counts are now indepentent of
the what buildings the current player happens to have.

Cleaned up the style of some of the code to be more maintainable.