Fixed a crash in the Hall of Fame screen from the original BtS code.
Previously the game would randomly crash when switching between
different filters in the Hall of Fame if there were a lot of replays.
For example, if there were lots of single player replays, and lots of
multiplayer replays, the game would often crash when switching between
single player view and multiplayer view.