Added a lot more nuance to the AI's evaluation of gifted units.
Gifts of great people will now always result in +1 for "you gave us
help".

Gifts of military units are now evaluated based on the usefulness of the
gifted unit, and the AI player's current military needs. (In general,
units won't be worth very much if the AI is not at war; and they'll be
worth a lot if the AI is losing a war.)