Rescaled and adjusted AI_warDiplomacyCost.
The previous version had some strange scaling based on population, which
had some counter-intuitive effects. (I'm not sure why I'd used that
scaling, and I'd even commented in the code that it was strange.) The
new scaling is based on the default number of players for the map size.

I've also made some adjustments to the weight. In general, the effect of
the team's "peace weight" has been dramatically reduced. (I did some
rough calculations and decided that the effect of peace weight was...
huge.)

I'd like some case data to get a sense of how well this function is
working, but I don't really want to test it right now...