Fixed a BtS pathfinding bug which caused inaccurate path turns.
Previously, if units with different speeds were grouped together moved with different movement costs, the pathfinder would fail to calculate the correct number of path turns. The most obvious example is when moving a group with mixed speeds along a railroad.

The bug was caused by the fact that the path nodes only store the lowest
moves value for the group, and so if the unit with more moves also has a
greater move cost, then the pathfinder will effectively be using the
lower moves with the higher movement cost.

I've fixed this by recalcuating the exact number of move for each unit
for the final turn of the path whenever there are mixed movement costs
involved. I expect this to have negligible speed inpact.

--

Also, I've changed CvSelectionGroup::maxMoves to be the minimum of the
group's max moves rather than the maximum. This seems far more sensible
to me.