Reversed the effect of bPrereqReligion for units and buildings.
Previously, if bPrereqReligion actually meant that a religion must _not_
be present in order to construct the building/unit. This is
counter-intuitive, and so I'd guess it was a mistake. bPrereqReligion
now means that a religion is required. (Thanks to sjodster for pointing
this out.)

Note: this change has no effect with the current xml, because
bPrereqReligion is zero for all units and buildings anyway.