Added some debug logging to CvGame::normalizeAddExtras
Features and bonuses added by this function will now be reported in the
bbai log file. (That file may not be ideal for this sort of information,
but it's convenient for me.)

I've also made some stylistic changes to the function.

BetterBTSAI.h now defines gMapLogLevel to control the new logging.