Fixed an int overflow problem in CvPlayerAI::AI_foundValue_bulk
For starting locations, the original BtS code included a large factor
for precision, which very frequently caused the int value to overflow -
meaning that starting locations were often evaluated incorrectly.

I've removed this factor, and added some additional checks at the end of
the function to (hopefully) detect similar problems in the future.