Made changes & fixes to the evaluation of starting positions.
Fixed a couple of problems in the original BtS handling of starting
positions which prevented the foundValue function from correctly
identifying existing starting positions. (eg. getMinOriginalStartDist
did not ignore it's own plot; and isStartingPlot did not count
dynamically chosen starting plots.)

Added some factors to the starting position evaluation to adjust for the
number of resources in the area, and the number of players already
placed in the area. The goal is to encourage players to be placed on the
most bountiful areas, and to reduce the likelyhood of isolated starts.

Fixed a bug which caused starting resource to be valued more highly than
intended.

Made a couple of other tweaks the the starting position evaluation.

Deleted some whitespace.