Made some large style adjustments to CvCityAI::AI_buildingValue
I've moved / changed some of the conditions in and around the main value
loop so that the code is easier to read / understand / maintain.

Also, BUILDINGFOCUS_WORLDWONDER is now defined such that it implies
BUILDINGFOCUS_WONDEROK.

A lot of the original code (which was commented out) has been deleted,
and some whitespace has been tided up.

Overall, the functionality should be roughly the same.