Tightened the conditions on the AI's "opportunistic wonder" builds.
'opportunistic wonder build 1' now requires that the player either has a
spare settler or a second city, and that they are not at war.

'opportunistic wonder build 2' now requires that the wonder be at least
as valuable to the city as whatever building the city would have built
instead of the wonder, also, the maximum build time is decreased when at
war.

Significantly reduced the chance of 'opportunistic wonder build 3' when
using 'total war', and when losing a war.