Changed some conditions and evaluations for bChop in AI_updateBestBuild
The bChop conditions now include a new case which allows new cities with
low productivity to chop regardless of how big the civ is, and
regardless of finacial trouble. The minimum city production and size for
this condition depends on leader personality.

Production from chopping is now valued more highly in cities with low
productivity.