Adjusted CvPlayerAI::AI_maxGoldTrade to better deal with high incomes.
AI_maxGoldTrade now has some dependance on the player's incoming and the
player's stockpiled gold, independant of how much gold the player has
traded in the past. This will fix the problem of very rich AI players
refusing to trade their gold when they have no other use for it.