Multi-completion of units will now be reflected by the city's order queue.
In CvCity::popOrder, if more than one unit is finished in a single turn
(due to production overflow), then the city will remove that same number
of queued units of the same type from the queue, regardless of the order
of the queue.

Also, tidied up some whitespace and comment.