Fixed a bug which broke CvCityAI::AI_chooseBuilding completely.
There was an order of precidence problem which was causing this function
to essentially choose the building for any non-zero odds.

This bug was introduced in 61167cfe7b75824003, when I tried to "Reduced
the effect of pre-existing progress in AI_chooseBuilding." That's a long
time ago for such a serious bug. I'm surprised I'm only noticing it
now...