Moved some of the AI_buildingValue veto conditions to fix a cache bug.
There are some conditions which cause AI_buildingValue to return zero.
However, in certain situations these conditions were being bypassed when
the building value cache was already set.

For example, if the city evaluated a non-wonder building and cached the
value, then later on tried to evaluate the building again with the
BUILDINGFOCUS_WORLDWONDER flag, the flag would be ignored because the
cached value would be used before the flag was checked.

This problem is now fixed. (But I'm also going to fix it in another way
in the next commit.)