Removed the temporary compatibilty hack for the 'disorganized' promotion
When the new promotion was added, I made some ad-hoc changes to the way
save games are loaded so that the new promotion didn't break old save
games.

This change was always intended to be temporary, and I've now removed
it. Hence, the next version of K-Mod will no longer be able to load save
games from before the promotion was added.

The temporary compatibility hack worked fine for K-Mod itself, but it
could cause problems for other mods which have different xml and so on.
In general, the DLL should not have to worry about the number of
promotions in the xml. The compatibility fix was just a kludge. Perhaps
in some later version I'll try to improve the save game format so that
such kludges are unnecessary.