Fixed a bug which caused the declare war move popup to appear twice.
When the player agrees to declare war, the war declaration and the move
command are both issued as net commands. Since the declare war popup is
now triggered when the net command is being sent - this makes the popup
trigger again before the war declaration is executed.

To fix this, I'm using the MOVE_DECLARE_WAR flag to indicate that we
shouldn't check for the popup. It's an ugly fix, and I'm sorry about
that.