Made some insignificant tweaks to worker AI for airbases and canals.
There was a bug on the final line of CvPlayerAI::AI_getPlotCanalValue
which effectively cancelled out the rest of the function. I've fixed
that bug, but I've also explicitly disabled the effect of canal value in
CvUnitAI::AI_workerMove anyway, because I don't think it's currently a
smart evaluation.

Also, I've tweaked AI_getPlotCanalValue and AI_getPlotAirbaseValue so
that the evaluation is not canceled by best-builds which do not affect
plot improvement.