Fixed another bug in the bbai danger cache, and made it more flexible.
In the bbai code, 'ActivePlayerNoDangerCache' was being set true if
there was no danger found regardless of the range being checked; and yet
this cache was being used for iRange <= DANGER_RANGE. This is a bug. For
example, the NoDangeCache might be set true after confirming that there
is no danger for iRange == 1, and then that cache could be used to
declare the plot safe for iRange == 4 regardless of what enemies might
be in that extended range.

The obvious way to fix this bug is to only set the cache to true if
iRange >= DANGER_RANGE, but I figure this would essentially remove any
performance gains from the cache - so instead I've rewritten it to be a
flexible safe range.

ActivePlayerNoDangerCache has been replaced with
ActivePlayerSafeRangeCache which signals the maximum range that is known
to be safe.