AI war evaluations can now handle defensive evaluations.
If we're already at war ('we' meaning *this in CvTeamAI), or if the
warplan is WARPLAN_ATTACKED or WARPLAN_ATTACKED_RECENT, then
AI_startWarVal and AI_warCommitmentCost will assume that we won't be the
ones declaring war and thus defensive pacts and diplomatic penalties are
not counted.

Removed the obsolete code for previous AI_startWarVal implementation.